move ImageViewer and subcomponents
