+++ src/components/index.ts
@@ -34,10 +34,7 @@ export { ImageSourceViewerProofreader } from './ImageSourceViewerProofreader';
 export { ImageSourceViewerTranslator } from './ImageSourceViewerTranslator';
 export { ImageTranslatorSettingHotKey } from './ImageTranslatorSettingHotKey';
 export { ImageTranslatorSettingMouse } from './ImageTranslatorSettingMouse';
-export { ImageViewer } from './ImageViewer';
-export { ImageViewerLabelTextExample } from './ImageViewerLabelTextExample';
-export { ImageViewerSettingPanel } from './ImageViewerSettingPanel';
-export { ImageViewerZoomPanel } from './ImageViewerZoomPanel';
+export { ImageViewer } from './project-file/ImageViewer';
 export { InvitationList } from './InvitationList';
 export { InviteUser } from './InviteUser';
 export { Label } from './Label';
@@ -49,10 +46,6 @@ export { ListSkeletonItem } from './ListSkeletonItem';
 export { LoadingIcon } from './LoadingIcon';
 export { MemberList } from './MemberList';
 export { MovableArea, MovableItem } from './Movable';
-export { MovableAreaColorBackground } from './MovableAreaColorBackground';
-export { MovableAreaImageBackground } from './MovableAreaImageBackground';
-export { MovableItemBars } from './MovableItemBars';
-export { MovableLabel } from './MovableLabel';
 export { NavTab } from './NavTab';
 export { NavTabs } from './NavTabs';
 export { Output } from './Output';
