+++ src/components/project-file/MovableItemBars.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/core';
-import { Icon } from '.';
+import { Icon } from '@/components';
 import React from 'react';
-import { FC } from '../interfaces';
+import { FC } from '@/interfaces';
 
 /** 用于拖动元素的三条线样的把手的属性接口 */
 interface MovableItemBarsProps {
